+++ src/Backend.php
@@ -1,15 +1,5 @@
 <?php
-/**
- * @brief myUrlHandlers, a plugin for Dotclear 2
- *
- * @package Dotclear
- * @subpackage Plugin
- *
- * @author Alex Pirine and contributors
- *
- * @copyright Jean-Christian Denis
- * @copyright GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html
- */
+
 declare(strict_types=1);
 
 namespace Dotclear\Plugin\myUrlHandlers;
@@ -19,12 +9,11 @@ use Dotclear\Core\Process;
 use Dotclear\Core\Backend\Favorites;
 
 /**
- * @brief   myUrlHandlers backend class.
- * @ingroup myUrlHandlers
+ * @brief       myUrlHandlers backend class.
+ * @ingroup     myUrlHandlers
  *
- * @author      Alex Pirine and contributors
- * @author      Jean-Christian Denis
- * @copyright   Alex Pirine
+ * @author      Alex Pirine (author)
+ * @author      Jean-Christian Denis (latest)
  * @copyright   GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html
  */
 class Backend extends Process
@@ -49,7 +38,7 @@ class Backend extends Process
                 'url'         => My::manageUrl(),
                 'small-icon'  => My::icons(),
                 'large-icon'  => My::icons(),
-                'permissions' => App::auth()->makePermissions([App::auth()::PERMISSION_CONTENT_ADMIN]),
+                'permissions' => App::auth()->makePermissions([App::auth()::PERMISSION_ADMIN]),
             ]);
         });